+++ R/vetiver-package.R
@@ -22,7 +22,8 @@ generics::augment
 #' @export
 generics::required_pkgs
 
-globalVariables(c("pr", ".metric", ".pred", "price", "tidy", "term", "estimate"))
+globalVariables(c("pr", ".metric", ".pred", "price", "tidy",
+                  "term", "estimate", "terms"))
 
 ## to avoid NOTE about "All declared Imports should be used."
 rapidoc_function_for_note <- function() {