+++ tests/testthat/test-create-ptype.R
@@ -23,6 +23,14 @@ test_that("ptype = FALSE", {
     )
 })
 
+test_that("ptype for model with interactions", {
+    cars_interaction <- lm(mpg ~ cyl * vs + disp, data = mtcars)
+    expect_equal(
+        vetiver_create_ptype(cars_interaction, TRUE),
+        vctrs::vec_slice(tibble::as_tibble(mtcars[, c(2, 8, 3)]), 0)
+    )
+})
+
 test_that("custom ptype", {
     expect_equal(
         vetiver_create_ptype(cars_lm, mtcars[3:10, 2:3]),